Prevent useless computation in SweepDetector through a tolerance
Sometimes `ruptures` does not return the specified number of change
points if no suitable change point is detected. Likewise the change in
cost can be miniscule, and by defination of the algorithms the change in
reduction in cost for adding an additional change point is less than the
current one (or at least for approximate solutions it is very unlikely
to be higher). Thus, we prevent the over penalization of high
`max_change_points` by stopping calculation if we don't have the correct
number of change points or the change in cost is below the tolerance.