Completely rework data pipeline model.
Now we use a composition approach which uses wrapping data generators
with data maps and data reducers. This is also composable meaning that
data maps can wrap data maps. This greatly simplifies the code and will
enable much higher performance with things like preventing the
duplication of the computation of a neighbor list.