Remove interactive call signature analysis.
The requirements to make it work are likely more expensive than the
feature is useful. A new function object must be created for each
instance to allow for the wrapping at every level and even then the
notebook introspections first like shows the original signature and the
line below it shows the wrapped one with the type annotations if
provided. However, since we cannot easily support documentation changes
due to this (without substantially more effort) even that has limited
utility.